+++ src/Addons/34thPRC_ArmourStandard/data/marines2/config_vehicles.hpp
@@ -113,6 +113,114 @@ class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Heavy_NoCollar: MA_Marine_He
 		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
 	};
 };
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Light: MA_Marine_Light_BDU
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Light";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Medium: MA_Marine_Medium_BDU
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Medium";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Heavy: MA_Marine_Heavy_BDU
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Heavy";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Light_NoCollar: MA_Marine_Light_BDU_NC
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Light_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Medium_NoCollar: MA_Marine_Medium_BDU_NC
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Medium_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
+class 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Blue_Heavy_NoCollar: MA_Marine_Heavy_BDU_NC
+{
+	scope=1;
+	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Urban_Blue_Heavy_NoCollar";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_UpperBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_LowerBDU.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Blue_TrooperCollar.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperSoftpadding_CO.paa",
+		"MA_Armor\data\Uniforms\Marine\data\Color_Variants\TrooperArmorStraps_CO.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_UpperArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_LowerArmor.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa",
+		"34thPRC_ArmourStandard\data\marines2\urban\Base_TrooperShoulders.paa"
+	};
+};
 class 34thPRC_ArmourStandard_Marines2_Trooper_Arctic_Light: 34thPRC_ArmourStandard_Marines2_Trooper_Urban_Light
 {
 	UniformClass="34thPRC_ArmourStandard_Marines2_Uniform_Arctic_Light";